refactor: separate performance_metrics into metrics module to resolve circular import
Address TODO in prophet/plot.py by extracting pure evaluation metrics to prophet/metrics.py. Re-export all symbols from prophet/diagnostics.py for backward compatibility. Add dedicated test suite in prophet/tests/test_metrics.py.
